Cambio de nombre del servicio localStatusService. Reordenamiento de lógica para no borrar siempre el currentEvent, modificación de nombres, modularización de método para calular si la sala debe quedar ocupada o no.
