+++ src/jobs/syncLocalResources.ts
@@ -1,5 +1,5 @@
 import { JobLocal } from "../schedulers/cronSetup";
-import localStatusService from "../services/localSyncService";
+import localStatusService from "../services/localStatusService";
 
 // Job local dedicado a limpiar y actualizar el estado de las salas según los eventos activos.
 export class SyncLocalResourcesJob implements JobLocal {
